third_party/WebKit/LayoutTests/fast/events/pointerevents/touch-pointer-events-expected.txt

Verifies that tapping on an element fires appropriate pointer events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


--- test PE+TE event sequences without preventDefault ---
green received pointerover at phase=target
lightgreen received pointerover at phase=bubbling
grey received pointerover at phase=bubbling
grey received pointerenter at phase=target
lightgreen received pointerenter at phase=target
green received pointerenter at phase=target
green received pointerdown at phase=target
lightgreen received pointerdown at phase=bubbling
grey received pointerdown at phase=bubbling
green received touchstart at phase=target
lightgreen received touchstart at phase=bubbling
grey received touchstart at phase=bubbling
green received pointermove at phase=target
lightgreen received pointermove at phase=bubbling
grey received pointermove at phase=bubbling
green received touchmove at phase=target
lightgreen received touchmove at phase=bubbling
grey received touchmove at phase=bubbling
green received pointerup at phase=target
lightgreen received pointerup at phase=bubbling
grey received pointerup at phase=bubbling
green received pointerout at phase=target
lightgreen received pointerout at phase=bubbling
grey received pointerout at phase=bubbling
green received pointerleave at phase=target
lightgreen received pointerleave at phase=target
grey received pointerleave at phase=target
green received touchend at phase=target
lightgreen received touchend at phase=bubbling
grey received touchend at phase=bubbling

--- test PE+TE event sequences with PE preventDefault ---
green received pointerover at phase=target
lightgreen received pointerover at phase=bubbling
grey received pointerover at phase=bubbling
grey received pointerenter at phase=target
lightgreen received pointerenter at phase=target
green received pointerenter at phase=target
green received pointerdown at phase=target
lightgreen received pointerdown at phase=bubbling
grey received pointerdown at phase=bubbling
green received touchstart at phase=target
lightgreen received touchstart at phase=bubbling
grey received touchstart at phase=bubbling
green received pointermove at phase=target
lightgreen received pointermove at phase=bubbling
grey received pointermove at phase=bubbling
green received touchmove at phase=target
lightgreen received touchmove at phase=bubbling
grey received touchmove at phase=bubbling
green received pointerup at phase=target
lightgreen received pointerup at phase=bubbling
grey received pointerup at phase=bubbling
green received pointerout at phase=target
lightgreen received pointerout at phase=bubbling
grey received pointerout at phase=bubbling
green received pointerleave at phase=target
lightgreen received pointerleave at phase=target
grey received pointerleave at phase=target
green received touchend at phase=target
lightgreen received touchend at phase=bubbling
grey received touchend at phase=bubbling

--- test PE params for single-touch ---
green received pointerover at phase=target
grey received pointerenter at phase=target
lightgreen received pointerenter at phase=target
green received pointerenter at phase=target
green received pointerdown at phase=target
PASS receivedPEsAtTarget[0].isPrimary is true
PASS receivedPEsAtTarget[0].button is 0
PASS receivedPEsAtTarget[0].buttons is 1
PASS receivedPEsAtTarget[0].clientX is 148
PASS receivedPEsAtTarget[0].clientY is 156
PASS receivedPEsAtTarget[1].isPrimary is true
PASS receivedPEsAtTarget[1].button is 0
PASS receivedPEsAtTarget[1].buttons is 1
PASS receivedPEsAtTarget[1].clientX is 148
PASS receivedPEsAtTarget[1].clientY is 156
PASS receivedPEsAtTarget[2].isPrimary is true
PASS receivedPEsAtTarget[2].button is 0
PASS receivedPEsAtTarget[2].buttons is 1
PASS receivedPEsAtTarget[2].clientX is 148
PASS receivedPEsAtTarget[2].clientY is 156
PASS receivedPEsAtTarget[3].isPrimary is true
PASS receivedPEsAtTarget[3].button is 0
PASS receivedPEsAtTarget[3].buttons is 1
PASS receivedPEsAtTarget[3].clientX is 148
PASS receivedPEsAtTarget[3].clientY is 156
PASS receivedPEsAtTarget[4].isPrimary is true
PASS receivedPEsAtTarget[4].button is 0
PASS receivedPEsAtTarget[4].buttons is 1
PASS receivedPEsAtTarget[4].clientX is 148
PASS receivedPEsAtTarget[4].clientY is 156
PASS receivedPEsAtTarget.length is 5
green received pointermove at phase=target
PASS receivedPEsAtTarget[5].isPrimary is true
PASS receivedPEsAtTarget[5].button is -1
PASS receivedPEsAtTarget[5].buttons is 1
PASS receivedPEsAtTarget[5].clientX is 168.5
PASS receivedPEsAtTarget[5].clientY is 175.25
PASS receivedPEsAtTarget[5].screenX is 168.5
PASS receivedPEsAtTarget[5].screenY is 175.25
PASS receivedPEsAtTarget[5].pageX is 168.5
PASS receivedPEsAtTarget[5].pageY is 175.25
PASS receivedPEsAtTarget.length is 6
green received pointerup at phase=target
green received pointerout at phase=target
green received pointerleave at phase=target
lightgreen received pointerleave at phase=target
grey received pointerleave at phase=target
PASS receivedPEsAtTarget[6].isPrimary is true
PASS receivedPEsAtTarget[6].button is 0
PASS receivedPEsAtTarget[6].buttons is 0
PASS receivedPEsAtTarget[7].isPrimary is true
PASS receivedPEsAtTarget[7].button is 0
PASS receivedPEsAtTarget[7].buttons is 0
PASS receivedPEsAtTarget[8].isPrimary is true
PASS receivedPEsAtTarget[8].button is 0
PASS receivedPEsAtTarget[8].buttons is 0
PASS receivedPEsAtTarget[9].isPrimary is true
PASS receivedPEsAtTarget[9].button is 0
PASS receivedPEsAtTarget[9].buttons is 0
PASS receivedPEsAtTarget[10].isPrimary is true
PASS receivedPEsAtTarget[10].button is 0
PASS receivedPEsAtTarget[10].buttons is 0
PASS receivedPEsAtTarget.length is 11

--- test PE isPrimary values ---
green received pointerover at phase=target
grey received pointerenter at phase=target
lightgreen received pointerenter at phase=target
green received pointerenter at phase=target
green received pointerdown at phase=target
PASS receivedPEsAtTarget[0].isPrimary is true
receivedPEsAtTarget[0].pointerId=5
PASS receivedPEsAtTarget[1].isPrimary is true
receivedPEsAtTarget[1].pointerId=5
PASS receivedPEsAtTarget[2].isPrimary is true
receivedPEsAtTarget[2].pointerId=5
PASS receivedPEsAtTarget[3].isPrimary is true
receivedPEsAtTarget[3].pointerId=5
PASS receivedPEsAtTarget[4].isPrimary is true
receivedPEsAtTarget[4].pointerId=5
green received pointerover at phase=target
grey received pointerenter at phase=target
lightgreen received pointerenter at phase=target
green received pointerenter at phase=target
green received pointerdown at phase=target
PASS receivedPEsAtTarget[5].isPrimary is false
receivedPEsAtTarget[5].pointerId=6
PASS receivedPEsAtTarget[6].isPrimary is false
receivedPEsAtTarget[6].pointerId=6
PASS receivedPEsAtTarget[7].isPrimary is false
receivedPEsAtTarget[7].pointerId=6
PASS receivedPEsAtTarget[8].isPrimary is false
receivedPEsAtTarget[8].pointerId=6
PASS receivedPEsAtTarget[9].isPrimary is false
receivedPEsAtTarget[9].pointerId=6
green received pointermove at phase=target
PASS receivedPEsAtTarget[10].isPrimary is false
PASS receivedPEsAtTarget[10].pointerId is 6
green received pointerup at phase=target
green received pointerout at phase=target
green received pointerleave at phase=target
lightgreen received pointerleave at phase=target
grey received pointerleave at phase=target
PASS receivedPEsAtTarget[11].isPrimary is true
PASS receivedPEsAtTarget[11].pointerId is 5
PASS receivedPEsAtTarget[12].isPrimary is true
PASS receivedPEsAtTarget[12].pointerId is 5
PASS receivedPEsAtTarget[13].isPrimary is true
PASS receivedPEsAtTarget[13].pointerId is 5
PASS receivedPEsAtTarget[14].isPrimary is true
PASS receivedPEsAtTarget[14].pointerId is 5
PASS receivedPEsAtTarget[15].isPrimary is true
PASS receivedPEsAtTarget[15].pointerId is 5
green received pointermove at phase=target
PASS receivedPEsAtTarget[16].isPrimary is false
PASS receivedPEsAtTarget[16].pointerId is 6
green received pointerover at phase=target
grey received pointerenter at phase=target
lightgreen received pointerenter at phase=target
green received pointerenter at phase=target
green received pointerdown at phase=target
PASS receivedPEsAtTarget[17].isPrimary is false
receivedPEsAtTarget[17].pointerId=7
PASS receivedPEsAtTarget[18].isPrimary is false
receivedPEsAtTarget[18].pointerId=7
PASS receivedPEsAtTarget[19].isPrimary is false
receivedPEsAtTarget[19].pointerId=7
PASS receivedPEsAtTarget[20].isPrimary is false
receivedPEsAtTarget[20].pointerId=7
PASS receivedPEsAtTarget[21].isPrimary is false
receivedPEsAtTarget[21].pointerId=7
green received pointerup at phase=target
green received pointerout at phase=target
green received pointerleave at phase=target
lightgreen received pointerleave at phase=target
grey received pointerleave at phase=target
PASS receivedPEsAtTarget[22].isPrimary is false
PASS receivedPEsAtTarget[22].pointerId is 6
PASS receivedPEsAtTarget[23].isPrimary is false
PASS receivedPEsAtTarget[23].pointerId is 6
PASS receivedPEsAtTarget[24].isPrimary is false
PASS receivedPEsAtTarget[24].pointerId is 6
PASS receivedPEsAtTarget[25].isPrimary is false
PASS receivedPEsAtTarget[25].pointerId is 6
PASS receivedPEsAtTarget[26].isPrimary is false
PASS receivedPEsAtTarget[26].pointerId is 6
green received pointerup at phase=target
green received pointerout at phase=target
green received pointerleave at phase=target
lightgreen received pointerleave at phase=target
grey received pointerleave at phase=target
PASS receivedPEsAtTarget[27].isPrimary is false
PASS receivedPEsAtTarget[27].pointerId is 7
PASS receivedPEsAtTarget[28].isPrimary is false
PASS receivedPEsAtTarget[28].pointerId is 7
PASS receivedPEsAtTarget[29].isPrimary is false
PASS receivedPEsAtTarget[29].pointerId is 7
PASS receivedPEsAtTarget[30].isPrimary is false
PASS receivedPEsAtTarget[30].pointerId is 7
PASS receivedPEsAtTarget[31].isPrimary is false
PASS receivedPEsAtTarget[31].pointerId is 7
green received pointerover at phase=target
grey received pointerenter at phase=target
lightgreen received pointerenter at phase=target
green received pointerenter at phase=target
green received pointerdown at phase=target
PASS receivedPEsAtTarget[32].isPrimary is true
PASS receivedPEsAtTarget[33].isPrimary is true
PASS receivedPEsAtTarget[34].isPrimary is true
PASS receivedPEsAtTarget[35].isPrimary is true
PASS receivedPEsAtTarget[36].isPrimary is true
PASS receivedPEsAtTarget.length is 37

PASS successfullyParsed is true

TEST COMPLETE
